detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.0.0
requests>=2.31.0
